refactor(api): package app assembly modules
