+++ src/qdash/api/app/router_registry.py
