+++ src/qdash/api/main.py
@@ -1,5 +1,5 @@
 """QDash API application entrypoint."""
 
-from qdash.api.app_factory import create_app
+from qdash.api.app import create_app
 
 app = create_app()
